+++ src/main/java/com/redbeemedia/enigma/core/player/timeline/BaseTimelineListener.java
@@ -1,5 +1,7 @@
 package com.redbeemedia.enigma.core.player.timeline;
 
+import com.google.android.exoplayer2.metadata.Metadata;
+import com.google.android.exoplayer2.source.hls.playlist.HlsMediaPlaylist;
 import com.redbeemedia.enigma.core.ads.AdEventType;
 import com.redbeemedia.enigma.core.ads.VastAdEntry;
 
@@ -27,6 +29,16 @@ public class BaseTimelineListener implements ITimelineListener {
 
     @Override
     public void onAdEvent(VastAdEntry entry, AdEventType eventType) {
+
+    }
+
+    @Override
+    public void onDashMetadata(Metadata metadata){
+
+    }
+
+    @Override
+    public void onHlsMetadata(HlsMediaPlaylist metadata){
 
     }
 }