+++ src/main/java/com/redbeemedia/enigma/core/player/timeline/ITimelineListener.java
@@ -1,5 +1,7 @@
 package com.redbeemedia.enigma.core.player.timeline;
 
+import com.google.android.exoplayer2.metadata.Metadata;
+import com.google.android.exoplayer2.source.hls.playlist.HlsMediaPlaylist;
 import com.redbeemedia.enigma.core.ads.AdEventType;
 import com.redbeemedia.enigma.core.ads.VastAdEntry;
 import com.redbeemedia.enigma.core.util.IInternalListener;
@@ -19,4 +21,6 @@ public interface ITimelineListener extends IInternalListener {
     void onBoundsChanged(ITimelinePosition start, ITimelinePosition end);
     void onLivePositionChanged(ITimelinePosition timelinePosition);
     void onAdEvent(VastAdEntry entry, AdEventType eventType);
+    void onDashMetadata(Metadata metadata);
+    void onHlsMetadata(HlsMediaPlaylist metadata);
 }