+++ src/test/java/com/redbeemedia/enigma/core/marker/MarkerPointsDetectorTest.java
@@ -4,6 +4,8 @@ import android.os.Handler;
 
 import androidx.annotation.NonNull;
 
+import com.google.android.exoplayer2.metadata.Metadata;
+import com.google.android.exoplayer2.source.hls.playlist.HlsMediaPlaylist;
 import com.redbeemedia.enigma.core.player.DefaultTimelinePositionFactoryTest;
 import com.redbeemedia.enigma.core.player.timeline.ITimeline;
 import com.redbeemedia.enigma.core.player.timeline.ITimelineListener;
@@ -141,6 +143,14 @@ public class MarkerPointsDetectorTest {
             @Override
             public boolean getVisibility() {
                 return false;
+            }
+
+            @Override
+            public void onDashMetadata(Metadata metadata) {
+            }
+
+            @Override
+            public void onHlsMetadata(HlsMediaPlaylist metadata) {
             }
         });
     }