+++ src/main/java/com/redbeemedia/enigma/core/player/IPlayerImplementation.java
@@ -1,5 +1,8 @@
 package com.redbeemedia.enigma.core.player;
 
+import android.content.Context;
+import android.support.v4.media.session.MediaSessionCompat;
+
 import com.google.android.exoplayer2.ui.PlayerNotificationManager;
 import com.google.android.exoplayer2.ui.SubtitleView;
 
@@ -7,6 +10,7 @@ public interface IPlayerImplementation {
     void install(IEnigmaPlayerEnvironment environment);
     void release();
     void setupPlayerNotificationManager(PlayerNotificationManager manager);
+    MediaSessionCompat createMediaSession(Context applicationContext);
     void updateTimeBar(long millis);
     SubtitleView getPlayerSubtitleView();
 }